requirements.txt
asgiref==3.8.1
azure-core==1.30.2
azure-storage-blob==12.21.0
backports.zoneinfo==0.2.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==4.2.14
django-storages==1.14.4
idna==3.7
isodate==0.6.1
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
urllib3==2.2.2
